# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JACS Studio API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/releases": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "List Releases",
        "operationId": "list_releases_api_v1_releases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReleaseResponse"
                  },
                  "type": "array",
                  "title": "Response List Releases Api V1 Releases Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "releases"
        ],
        "summary": "Create Release",
        "operationId": "create_release_api_v1_releases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/releases/{release_id}/publish": {
      "post": {
        "tags": [
          "releases"
        ],
        "summary": "Publish Release",
        "operationId": "publish_release_api_v1_releases__release_id__publish_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/releases/{release_id}/unpublish": {
      "post": {
        "tags": [
          "releases"
        ],
        "summary": "Unpublish Release",
        "operationId": "unpublish_release_api_v1_releases__release_id__unpublish_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/releases/{release_id}": {
      "delete": {
        "tags": [
          "releases"
        ],
        "summary": "Delete Release",
        "operationId": "delete_release_api_v1_releases__release_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "release_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Release Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/releases/check": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Check Update",
        "operationId": "check_update_api_v1_releases_check_get",
        "parameters": [
          {
            "name": "platform",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(windows|macos)$",
              "title": "Platform"
            }
          },
          {
            "name": "current_version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^v?\\d+(\\.\\d+)*.*$",
              "title": "Current Version"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(stable|beta)$",
              "default": "stable",
              "title": "Channel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReleaseChannel": {
        "type": "string",
        "enum": [
          "stable",
          "beta"
        ],
        "title": "ReleaseChannel"
      },
      "ReleaseCreate": {
        "properties": {
          "version": {
            "type": "string",
            "pattern": "^v\\d+\\.\\d+\\.\\d+$",
            "title": "Version"
          },
          "platform": {
            "type": "string",
            "pattern": "^(windows|macos)$",
            "title": "Platform"
          },
          "channel": {
            "$ref": "#/components/schemas/ReleaseChannel",
            "default": "stable"
          },
          "download_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Download Url"
          },
          "sha512": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{128}$",
            "title": "Sha512"
          },
          "release_notes": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Release Notes"
          },
          "signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature"
          },
          "force_update": {
            "type": "boolean",
            "title": "Force Update",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "version",
          "platform",
          "download_url",
          "sha512",
          "release_notes"
        ],
        "title": "ReleaseCreate"
      },
      "ReleaseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": "stable"
          },
          "download_url": {
            "type": "string",
            "title": "Download Url"
          },
          "sha512": {
            "type": "string",
            "title": "Sha512"
          },
          "release_notes": {
            "type": "string",
            "title": "Release Notes",
            "default": ""
          },
          "force_update": {
            "type": "boolean",
            "title": "Force Update",
            "default": false
          },
          "signature": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature"
          },
          "rollout_percent": {
            "type": "integer",
            "title": "Rollout Percent",
            "default": 100
          },
          "min_app_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min App Version"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "published"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version",
          "platform",
          "download_url",
          "sha512"
        ],
        "title": "ReleaseResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}